feat: config Docker and fix some ulrs
