+++ backend/docker-compose.yml
@@ -1,0 +1,29 @@
+# version: '3.8'
+
+services:
+  web:
+    build: .
+    command: python manage.py runserver 0.0.0.0:8000
+    volumes:
+      - app_data:/usr/src/app/
+    ports:
+      - "8000:8000"
+    env_file:
+      - ./env.dev
+    depends_on:
+      - db
+  
+  db:
+    image: postgres:13.0-alpine
+    ports:
+      - 5433:5432
+    volumes:
+      - postgres_data:/var/lib/postgresql/data
+    environment:
+      - POSTGRES_USER=twitter_clone
+      - POSTGRES_PASSWORD=twitter_clone
+      - POSTGRES_DB=twitter_clone_db
+
+volumes:
+  postgres_data:
+  app_data:
