+++ backend/env.dev
@@ -1,0 +1,9 @@
+DEBUG=1
+SECRET_KEY=foo
+DJANGO_ALLOWED_HOSTS=localhost,127.0.0.1,[::1]
+SQL_ENGINE=django.db.backends.postgresql
+SQL_DATABASE=twitter_clone_db
+SQL_USER=twitter_clone
+SQL_PASSWORD=twitter_clone
+SQL_HOST=db
+SQL_PORT=5432
