+++ backend/pyproject.toml
@@ -5,6 +5,7 @@ description = "twitter_clone"
 authors = ["Diego Cavalcanti <cavalcantidiego@hotmail.com>"]
 readme = "README.md"
 
+
 [tool.poetry.dependencies]
 python = "^3.13"
 django = "^5.1.4"
@@ -17,6 +18,8 @@ python-decouple = "^3.8"
 factory-boy = "^3.3.1"
 pytest = "^8.3.4"
 pillow = "^11.0.0"
+django-extensions = "^3.2.3"
+djangorestframework-simplejwt = "^5.3.1"
 
 
 [build-system]
