+++ backend/users/urls.py
@@ -1,8 +1,10 @@
 from django.urls import path, include
-from rest_framework.routers import DefaultRouter
+# from rest_framework.routers import DefaultRouter
+from rest_framework import routers
 from .views import UserViewSet
 
-router = DefaultRouter()
+# router = DefaultRouter()
+router = routers.SimpleRouter()
 router.register(r'', UserViewSet)
 
 urlpatterns = [
